Application/intouch/requirements.txt
absl-py==0.9.0
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
boto3==1.14.63
botocore==1.17.63
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
Django==3.0.8
django-filter==2.3.0
django-storages==1.10.1
djangorestframework==3.11.1
docutils==0.15.2
dominate==2.5.1
Faker==4.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.3
future==0.18.2
gast==0.3.3
gevent==20.6.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
greenlet==0.4.16
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
idna==2.10
image==1.5.32
imageio==2.9.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Keras==2.4.3
Keras-Applications==1.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
networkx==2.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
pandas==1.0.5
pdfkit==0.6.1
pefile==2019.4.18
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pylint==2.5.3
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
rdflib==5.0.0
reportlab==3.5.47
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.7.4
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.25
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.4
zope.event==4.4
zope.interface==5.1.0
